+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.1 under development
 
 - Enh #286: Change property `Schema::$typeMap` to constant `Schema::TYPE_MAP` (@Tigrov)
+- Bug #287: Fix `DMLQueryBuilder::insertWithReturningPks()` and `Command::insertWithReturningPks()` methods (@Tigrov)
 
 ## 1.1.0 November 12, 2023